requirements.txt
attrs==20.3.0
Bottleneck==1.3.2
coloredlogs==14.0
Cython==0.29.21
decorator==4.4.2
graphviz==0.14.2
humanfriendly==8.2
importlib-metadata==2.0.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
MarkupSafe==1.1.1
networkx==2.5
numexpr==2.7.1
numpy==1.19.4
pandas==1.1.4
pylsl==1.13.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-osc==1.7.4
pytz==2020.4
PyYAML==5.3.1
pyzmq==19.0.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
tables==3.6.1
threadpoolctl==2.1.0
timeflux==0.6.1
xarray==0.16.1
zipp==3.4.0
python-dotenv==0.15.0
.
